requirements.txt
numpy==1.26.3
pandas==2.2.0
Pillow==10.2.0
plotly==5.18.0
scikit_learn==1.3.2
streamlit==1.30.0
